Auth/use components (#60)
