+++ client/.env-sample
@@ -1,3 +1,4 @@
 VITE_API_URL_BASE= 
 VITE_CLOUD_PRESET=
 VITE_CLOUD_URL=
+VITE_CLERK_PUB_KEY
