+++ client/src/pages/Auth/Layout.jsx
@@ -1,0 +1,9 @@
+import { Outlet } from 'react-router-dom'
+
+export function AuthLayout() {
+  return (
+    <div className="flex min-h-screen items-center justify-center bg-hint-of-red">
+      <Outlet />
+    </div>
+  )
+}
